refactor(view): View 분리
기존 Screen 하나의 View 분리
- Common, Input, Output으로 파일 및 함수 분리
