+++ src/main/java/racingcar/view/Common.java
@@ -1,0 +1,13 @@
+package racingcar.view;
+
+public class Common {
+    public static final String EMPTY_STRING = "";
+
+    public static void printEmptyLine() {
+        System.out.println(EMPTY_STRING);
+    }
+
+    public static void printError(String message) {
+        System.out.print(message);
+    }
+}
